Moved Daily Rashifal to Daily Panchang and corrected color scheme
